move all YASL libraries to same folder
