+++ std/yasl-std-io.c
